Changed slider's value_increment to do what it says
value_increment now controls in which increments the slider's value
moves. If value_min is 0, value_max is 100, and value value_increment
is 25, then the value will be either 0, 25, 50, 75, or 100.

Now, the thumb's size is calculated based on value_increment, so the
size of the thumb gives a visual indication of value increment.