Added "text_format" to sliders
If the user wants to print the value of a slider, the text_format
property can be set to a string with standard printf formatting.
However, the only valid values are %f and %p, where %p is used to show
the slider's value as a percentage.

For example, the following would show "Value: 25":
    {
        type = "slider
        , value = 25
        , value_min = 0
        , value_max = 50
        , value_increment = 1
        , text_format = "Value: %f"
    }
and the following would show "Value: 50%":
    {
        type = "slider
        , value = 25
        , value_min = 0
        , value_max = 50
        , value_increment = 1
        , text_format = "Value: %p%%"
    }

Simple error checking is performed, and the text will print INVALID if
the text string is invalid. It must contain one and only one instance of
either %p or %f to be considered valid.

Max length is currently set to 255 characters, which I feel should be
enough for most applications.